Docs updates for the new wrapper API! [ci skip]
